Move the event objects to src/internal/events.ts
The Event classes, DataTransfer, EventTarget with its listener store,
the event handler attributes, and the helpers that go with them leave
dom.ts. events.ts imports nothing from dom.ts at runtime: it finds a
target's document through the public node and window shape, reports
errors through the window's own reportError, and leaves dispatchEvent to
dom.ts, which installs it on the prototype beside the dispatch algorithm
it belongs to. The get-the-parent protocol becomes one function in
dom.ts over the tree types instead of a symbol-keyed method four classes
overrode. dom.ts imports 57 names back, and the drift check reads the
event classes from the new module.
